Fix initialisation defect in the Item Editor.
Before this fix, the following steps could reliably be followed to
reproduce the defect:
1. Load the list
2. Click 'Edit' on the first item.
3. Delete everything in the 'name' field.
4. Click outside the modal dialog. (This dismisses it.)
5. Click 'Edit' again. The name field is now empty.
6. Click 'Save'
There is possibly more than one defect here. The most significant defect
is that it is possible to save a list item with an empty name. This is
not supposed to be possible.
The fix is to use an object to track which value should be used to
initialise the form control when the modal dialog is shown.